# Optimized Unix2DosFilter to use Span<T>.IndexOf() and CopyTo()